+++ src/test/java/com/diffplug/common/rx/RxExample.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 DiffPlug
+ * Copyright (C) 2020-2022 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,20 +16,20 @@
 package com.diffplug.common.rx;
 
 
-import com.diffplug.common.collect.ImmutableSet;
-import com.diffplug.common.collect.Immutables;
-import io.reactivex.Observable;
 import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.Graphics;
 import java.awt.Point;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
+import java.util.HashSet;
+import java.util.LinkedHashSet;
 import java.util.Optional;
 import java.util.Set;
 import javax.swing.JComponent;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
+import kotlinx.coroutines.flow.FlowKt;
 
 @SuppressWarnings("serial")
 public class RxExample extends JFrame {
@@ -57,7 +57,7 @@ public class RxExample extends JFrame {
 		/** The cell which the mouse is over. */
 		private RxGetter<Optional<Integer>> rxMouseOver;
 		/** The selected cells. */
-		private RxBox<ImmutableSet<Integer>> rxSelection;
+		private RxBox<Set<Integer>> rxSelection;
 
 		RxGrid() {
 			// maintain the position of the mouse
@@ -81,12 +81,13 @@ public class RxExample extends JFrame {
 			});
 
 			// maintain the selection state
-			rxSelection = RxBox.of(ImmutableSet.of());
+			rxSelection = RxBox.of(new HashSet<>());
 			addMouseListener(new MouseAdapter() {
 				@Override
 				public void mouseClicked(MouseEvent e) {
 					rxMouseOver.get().ifPresent(cell -> {
-						rxSelection.modify(Immutables.mutatorSet(selection -> {
+						rxSelection.modify(set -> {
+							HashSet<Integer> selection = new LinkedHashSet<>(set);
 							if (e.isControlDown()) {
 								// control => toggle mouseOver item in selection
 								if (selection.contains(cell)) {
@@ -99,13 +100,14 @@ public class RxExample extends JFrame {
 								selection.clear();
 								selection.add(cell);
 							}
-						}));
+							return selection;
+						});
 					});
 				}
 			});
 
 			// trigger a repaint on any change
-			Rx.subscribe(Observable.merge(rxMouseOver.asObservable(), rxSelection.asObservable()), anyChange -> {
+			Rx.subscribe(FlowKt.merge(rxMouseOver.asObservable(), rxSelection.asObservable()), anyChange -> {
 				this.repaint();
 			});
 		}