+++ src/test/java/com/diffplug/common/rx/RxOrderedSetTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 DiffPlug
+ * Copyright (C) 2020-2022 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,8 +16,8 @@
 package com.diffplug.common.rx;
 
 
-import com.diffplug.common.collect.ImmutableList;
 import com.diffplug.common.rx.RxOrderedSet.OnDuplicate;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import org.junit.Assert;
@@ -38,9 +38,9 @@ public class RxOrderedSetTest {
 
 	private void testCase(List<Integer> before, List<Integer> after, OnDuplicate policy, List<Integer> expected) {
 		// create the initial list with its policy
-		RxOrderedSet<Integer> list = RxOrderedSet.of(ImmutableList.copyOf(before), policy);
+		RxOrderedSet<Integer> list = RxOrderedSet.of(new ArrayList<>(before), policy);
 		// set the new value
-		list.set(ImmutableList.copyOf(after));
+		list.set(new ArrayList<>(after));
 		// test the resulting value
 		Assert.assertEquals(expected, list.get());
 	}